Fix fault detection
Was reading wrong register for controller faults. Need to read 0x0122 (290) and not 0x0121 (289). Had no way of testing this until recently I got a panel overvoltage fault by accident. Also the wrong fault text was being returned by one and needed an offset of -1 when selecting it. Not sure why and what the ramifications are for the first and last faults in the list.